+++ src/cjs/utils/datastructures/Color.js
@@ -210,18 +210,37 @@ var Color = /** @class */ (function () {
         this.saturate("-" + v);
         return this;
     };
+    //   lighten(v: string | number): Color {
+    //     if ("string" == typeof v && v.indexOf("%") > -1 && (v = parseInt(v)) != NaN) {
+    //       this.l += v / 100;
+    //     } else if ("number" == typeof v) {
+    //       if (v >= -1.0 && v <= 1.0) {
+    //         // range 0.0...1.0
+    //         this.l += v;
+    //       } else {
+    //         // range 255
+    //         this.l += v / 255;
+    //       }
+    //     } else {
+    //       throw new Error("error: bad modifier format (percent or number)");
+    //     }
+    //     if (this.l > 1) this.l = 1;
+    //     else if (this.l < 0) this.l = 0;
+    //     Color.Converter.HSLToRGB(this);
+    //     return this;
+    //   }
     Color.prototype.lighten = function (v) {
         if ("string" == typeof v && v.indexOf("%") > -1 && (v = parseInt(v)) != NaN) {
-            this.l += v / 100;
+            this.l += (1 - this.l) * (v / 100);
         }
         else if ("number" == typeof v) {
             if (v >= -1.0 && v <= 1.0) {
                 // range 0.0...1.0
-                this.l += v;
+                this.l += (1 - this.l) * v;
             }
             else {
                 // range 255
-                this.l += v / 255;
+                this.l += (1 - this.l) * (v / 255);
             }
         }
         else {
@@ -234,8 +253,32 @@ var Color = /** @class */ (function () {
         Color.Converter.HSLToRGB(this);
         return this;
     };
+    //   darken(v: string | number): Color {
+    //     this.lighten(typeof v === "string" ? "-" + v : -v);
+    //     return this;
+    //   }
     Color.prototype.darken = function (v) {
-        this.lighten(typeof v === "string" ? "-" + v : -v);
+        if ("string" == typeof v && v.indexOf("%") > -1 && (v = parseInt(v)) != NaN) {
+            this.l -= this.l * (v / 100);
+        }
+        else if ("number" == typeof v) {
+            if (v >= -1.0 && v <= 1.0) {
+                // range 0.0...1.0
+                this.l -= this.l * v;
+            }
+            else {
+                // range 255
+                this.l -= this.l * (v / 255);
+            }
+        }
+        else {
+            throw new Error("error: bad modifier format (percent or number)");
+        }
+        if (this.l > 1)
+            this.l = 1;
+        else if (this.l < 0)
+            this.l = 0;
+        Color.Converter.HSLToRGB(this);
         return this;
     };
     Color.prototype.fadein = function (v) {